Moved shared code to hpp file
